pyproject.toml
[project]
name = "eda-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai-agents[litellm]>=0.0.15",
    "litellm[google]>=1.40.0",
    "streamlit>=1.35.0",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.13.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
